#13 화요일 Follow / 오류해결 못함 사이각 비교하고 방향회전을 멈춰야하는데 PI/90.f가 -100만쯤 나와서 무조건 회전하는 오류있음

+++ inho/inho/CPathMgr.cpp
@@ -9,9 +9,10 @@ void CPathMgr::init() {
 
 	GetCurrentDirectory(255, g_szContent);
 
-	size_t Len = wcslen(g_szContent);
+	size_t Len = wcslen(g_szContent); 
 
-	for (int i = Len - 1; i >= 0; --i) {
+	for (size_t i = Len - 1; i >= 0; --i) {
+		
 		if ('\\' == g_szContent[i]) {
 			g_szContent[i + 1] = '\0';
 			break;
